# stable-diffusion : ggml-alloc integration and gpu acceleration