--- a/.gitignore
+++ b/.gitignore
@@ -1,6 +1,6 @@
 build*/
 test/
-
+.vscode/
 .cache/
 *.swp
 .vscode/