--- a/.gitignore
+++ b/.gitignore
@@ -7,3 +7,4 @@ test/
 *.bat
 *.bin
 *.exe
+*.gguf