--- a/common/common.cpp
+++ b/common/common.cpp
@@ -2,9 +2,10 @@
 #include <string>
 #include "common.h"
 #include <thread>
-#include <unordered_set>
 #include <fstream>
-#include <string.h>
+#include <iostream>
+#include <cstring>
+#include <unordered_set>
 
 #if defined(__APPLE__) && defined(__MACH__)
 #include <sys/sysctl.h>