--- a/models/.gitignore
+++ b/models/.gitignore
@@ -1,5 +1,0 @@
-*.bin
-*.ckpt
-*.safetensor
-*.safetensors
-*.log