… CMakeLists.txt
… README.md
… common/CMakeLists.txt
… common/json.hpp
… common/miniz.h
… common/stb_image.h
… common/stb_image_write.h
… common/zip.c
… common/zip.h
… compare.exe
… examples/CMakeLists.txt
… examples/cli/CMakeLists.txt
… examples/convert/CMakeLists.txt
… examples/convert/README.md
… examples/convert/convert.cpp
… models/README.md
… models/convert.py
… models/convert_diffusers_to_original_stable_diffusion.py
… models/requirements.txt
… models/vocab.json
… output.png
… output_2.png
… output_3.png
… output_4.png
… stable-diffusion.cpp
… stable-diffusion.h