# Context
I had been waiting for a response #72 from @leejet regarding updating the GGML version so I could integrate `ggml-alloc`, the new memory management approach for GGML (also necessary for integrating the CUDA backend). He synchronized it with the latest changes in @ggerganov GGML repository, but it seems that in the process, the GGML library stopped working directly (at least on my computer, it's causing issues when computing) because it still has the dynamic mode integrated.

I'm using [this](https://github.com/FSSRepo/ggml/tree/master) ggml version in this PR.

I have removed the ggml submodule and directly implemented one that I already have, in which I refactored the Conv1D and Conv2D operations to use `ggml_im2col` and `ggml_mul_mat`. This will make it easier for me to create the different CUDA kernels.

## Build this PR

```bash
git clone https://github.com/FSSRepo/stable-diffusion.cpp.git
mkdir build
cd build
cmake ..
cmake --build . --config Release
```

I've noticed that `ggml-alloc` consumes much more memory than leejet's current `dynamic mode`. I'm likely overlooking some configuration to prevent extensive compute graphs from consuming too much memory with `ggml-alloc`.

# Here are the comparative details and results of the integration with `ggml-alloc`

```c++
./sd  -m kotosmix_q4_0.bin --verbose -p "beautiful anime girl, white hair, blue eyes, realistic, 4k, high quality" -n "bad quality, ugly" --steps 10 -t 6 -s 48747
```

**Note:** the stable-diffusion.cpp commit afec505 that worked for me.

Memory usage:

| compute graph | dynamic mode (runtime memory) | ggml-alloc (compute buffer size) |
| ------- | -------- | -------- |
| learned condition (CLIP Text Encoder)   | 13.37MB | 25.54 MB |
| diffusion (UNet Model) | 570.65MB   | 1147.33 MB |
| AutoEncoderKL (VAE Decode) | 1674.14MB | 2496.00 MB |

Timings:

| compute graph | dynamic mode | ggml-alloc |
| ------- | -------- | -------- |
| learned condition (CLIP Text Encoder)   | 0.26 s | 0.23 s |
| diffusion (UNet Model) | ~ 18.22 s per sample | ~ 13.81 s per sample |
| AutoEncoderKL (VAE Decode) | 27 s | 24.22 s |

Result image (the same in both cases):

![output](https://github.com/leejet/stable-diffusion.cpp/assets/57494570/1a748563-0ca9-4d8f-b139-b4dc3e712613)

Over the next few days, I will be making some improvements, trying to investigate and gather feedback with `ggml-alloc` to avoid the current memory usage, and once I have fixed that, I can start working on implementing the CUDA kernels.

# Progress overview

Here's a summary of what I will be carrying out in the coming days to complete this PR:

ggml operations that utilize stable diffusion and their availability across backends.:

| operation | CPU backend | CUDA backend | Kernel Status |
| ------- | -------- | -------- | -------- |
| ADD | ✅ | ✅ |  Support broadcasting to avoid `ggml_repeat` |
| MUL | ✅ | ✅ |  Support broadcasting to avoid `ggml_repeat` |
| CONCAT | ✅ | ❌ (need implement) | |
| NORM | ✅ | ✅ |
| GET_ROWS | ✅ | ✅ |
| DIAG_MASK_INF | ✅ | ✅ |
| GROUP_NORM | ✅ | ❌ (need implement)  | The currently implemented kernel is too inefficient. |
| MUL_MAT | ✅ | ✅ | |
| SCALE | ✅ | ✅ |
| PERMUTE | ✅ | - |
| RESHAPE | ✅ | - |
| VIEW | ✅ | - |
| SOFT_MAX | ✅ | ✅ |
| IM2COL |  ✅  | ✅ |
| UPSCALE | ✅  | ❌ (need implement)  |
| UNARY |  ✅  | ❌ (only GeLU and SiLU, missing GeLU Quick) |
| PAD | ❌  | ❌ | VAE decoder requires zero pad a tensor

Tasks for complete this PR:

- [x] Integrate ggml-alloc to simplify the use of ggml-backend.
- [x] Improve memory usage, determine if I'm using the `ggml-alloc` API correctly, I might be missing something that could help me reduce unnecessary memory usage during compute. can you help me? @slaren @ggerganov.
- [x] Implement the missing CUDA kernels `ggml_concat`, `ggml_group_norm`, `ggml_upscale`, `ggml_gelu_quick`.
- [x] Offload the model to the GPU and conduct performance and memory usage tests.
- [x] Implement the use of custom VAEs. I will try to implement the `gguf` format to simplify the loading of tensors.