@ggerganov 2023-10-23
> I've noticed that ggml-alloc consumes much more memory than leejet's current dynamic mode. I'm likely overlooking some configuration to prevent extensive compute graphs from consuming too much memory with ggml-alloc.

Nothing comes directly to mind. Need to have a more detailed look in the code when I get some time.
One possible error is to avoid reusing the same compute buffer for new graphs - I had some troubles with that when adding `ggml-alloc` to `whisper.cpp`: https://github.com/ggerganov/whisper.cpp/pull/1270#discussion_r1320847591
It does not seem very related the increased memory that you observed though.