@Jonathhhan 2023-10-23
@FSSRepo thanks.
I need to comment out line 245-257 in stable-diffusion.cpp, then it works for me:
```
    //if(allocr != NULL) {
        //ggml_allocr_alloc(allocr, embedding);
    //}
```
Otherwise I get: `GGML_ASSERT: ggml-alloc.c:116: tensor->data == 0`
One step takes about 11s with 512x512, which is faster than before.