@FSSRepo 2023-10-23
> Otherwise I get: `GGML_ASSERT: ggml-alloc.c:116: tensor->data == 0`

I will try to reproduce this, because it's a strange behaivor