@Jonathhhan 2023-10-23
> Can you provide more information, system operative, hardware

Of course: Windows 11 / VS / CUDA / GTX 3090 / i7-12700k