@FSSRepo 2023-10-23
> I need to comment out line 245-257 in stable-diffusion.cpp

> ```
>     //if(allocr != NULL) {
>         //ggml_allocr_alloc(allocr, embedding);
>     //}
> ```
245 to 257 are 12 lines or you refer 245 to 247?