@FSSRepo 2023-10-24
> With `/D GGML_USE_CUBLAS` one step 512x512 is between 8 and 9 seconds.

The backend to compute is fixed to cpu_backend