@FSSRepo 2023-10-24
> > The backend to compute is fixed to cpu_backend
> 
> I see. Interestingly it is already a little faster (8~ vs 10~ seconds/step).

Possibly matrix multiplications are being delegated to the GPU.