@FSSRepo 2023-10-24
> I need to comment out line 245-247 in stable-diffusion.cpp, then it works for me:

Are you using stable diffusion 2 when this error happen?