@FSSRepo 2023-10-24
> With the main branch I get this error: GGML_ASSERT: ggml.c: ctx->mem_buffer != NULL [imartinez/privateGPT#90](https://github.com/imartinez/privateGPT/issues/90) So maybe it is not enough space on my HDD...

What matter that in this context?