@Jonathhhan 2023-10-24
@FSSRepo yes, it was with stable diffusion 2. I had that error `GGML_ASSERT: ggml.c: ctx->mem_buffer != NULL` with the main branch, so I thought maybe this one is caused by missing disc space too (but no it is not, so it does not matter in this context).