@slaren 2023-10-24
Instead of calling `ggml_allocr_alloc` after every tensor, you can do something like this to allocate all the tensors created in a context:
```C++
for (struct ggml_tensor * t = ggml_get_first_tensor(ctx); t != NULL; t = ggml_get_next_tensor(ctx, t)) {
        ggml_allocr_alloc(alloc, t);
    }
```
You can also use a similar loop calling `ggml_nbytes` to calculate the size of the buffer.

Caveat: the functions to enumerate the tensors in a context are only in the llama.cpp repository currently, they still need to be synced to the ggml repo.

Related: https://github.com/ggerganov/ggml/issues/578