@leejet 2023-10-24
To resolve merge conflicts, I rebased the master branch of FSSRepo/stable-diffusion.cpp onto the latest master branch of leejet/stable-diffusion.cpp. I also updated the URL for GGML to https://github.com/FSSRepo/ggml/tree/master to make it easier to review the changes and facilitate future merges.