@leejet 2023-10-24
> It seems like I mistakenly closed this pull request.

I am unable to push my changes to https://github.com/FSSRepo/stable-diffusion.cpp/tree/master at the moment, which prevents me from reopening this pull request. Can you please pull https://github.com/leejet/stable-diffusion.cpp/tree/pr75 and push the corresponding changes to https://github.com/FSSRepo/stable-diffusion.cpp/tree/master? This way, I can reopen this pull request.