@FSSRepo 2023-10-24
> Sorry for the inconvenience

Don't worry. Anyway, I will need your help to understand how some aspects of this stable diffusion implementation work.