@FSSRepo 2023-10-24
> I still can't reopen this PR. It says 'there are now new commits.' It might require you to push the new changes to the [master](https://github.com/FSSRepo/stable-diffusion.cpp/tree/master).

I will try it