@leejet 2023-10-24
I've updated the submodule 'ggml' in this branch https://github.com/leejet/stable-diffusion.cpp/tree/pr75 to the latest commit of FSSRepo/sd-cpp-fixes, and this branch also includes the changes you made to stable-diffusion.cpp earlier.