@leejet 2023-10-24
You can pull this branch to your local machine, rebase your local stable-diffusion.cpp master branch to this branch, and then push the changes to https://github.com/FSSRepo/stable-diffusion.cpp/tree/master.