@FSSRepo 2023-10-24
@leejet I'm trying to get the kernel to work for the 'group_norm' operation, but I haven't been able to perform the sum in parallel, mainly due to the limit of concurrent threads and the size of the tensors passing through this operation (very large, and this operation in its CPU implementation seems to be not very parallelizable). I want to understand the difference between `ggml_norm` and `ggml_group_norm` and thus arrive at a clearer approach.