@leejet 2023-10-25
> I want to understand the difference between ggml_norm and ggml_group_norm and thus arrive at a clearer approach.

I think this might help

https://gaoxiangluo.github.io/2021/08/01/Group-Norm-Batch-Norm-Instance-Norm-which-is-better/figure6.png