@FSSRepo 2023-10-25
> > I want to understand the difference between ggml_norm and ggml_group_norm and thus arrive at a clearer approach.
> 
> I think this might help
> 
> https://gaoxiangluo.github.io/2021/08/01/Group-Norm-Batch-Norm-Instance-Norm-which-is-better/figure6.png

![Screenshot 2023-10-25 071918](https://github.com/leejet/stable-diffusion.cpp/assets/57494570/b65bcb9f-822c-46f1-8ac1-5334b0f94a7c)


Seems good. While I was showering, the solution came to me, hahaha.