@FSSRepo 2023-10-25
group norm kernel finished 😎

Tensor 3D [4, 4, 8] and n_group = 4 (left: CUDA, right: ggml cpu backend):

![Screenshot 2023-10-25 125414](https://github.com/leejet/stable-diffusion.cpp/assets/57494570/c94de215-8155-451b-abbc-6ce5ba2ce3f5)