@FSSRepo 2023-10-25
@slaren Is there a function to transfer GPU data to CPU without having to create this function? `tensor->data` in this case is a pointer of memory in gpu.

```c++
void ggml_fallback_tensor_f32(struct ggml_tensor* tensor) {
    if(ggml_backend_is_cpu(ggml_get_backend(tensor))) {
        return;
    }
    // for cuda backend bring data from device
    assert(tensor->type == GGML_TYPE_F32);
    int size_ = ggml_nbytes(tensor);
    float* data = (float*)malloc(size_);
    ggml_backend_tensor_get(tensor, data, 0, size_);
    tensor->data = data;
}
```