@slaren 2023-10-25
You can use `ggml_backend_tensor_copy` to copy a tensor between backends. Allocate a tensor on the CPU backend with the exact same shape and call `ggml_backend_tensor_copy(gpu_tensor, cpu_tensor)`. Is that what you need? You cannot just change the `data` pointer of a tensor.