@FSSRepo 2023-10-26
@leejet I've noticed that this is very sensitive to very small changes. Previously, the 'eps' parameter of 'ggml norm' was set to 1e-5. When you updated ggml, and you had to manually set that parameter, you changed it to 1e-6. This resulted in the appearance of small artifacts and variations in the final image, especially in the CLIP model.

![Sin título](https://github.com/leejet/stable-diffusion.cpp/assets/57494570/d8305f2e-9120-44e3-84f6-166e354ea7aa)